requirements.txt
Django==2.0
pytz==2019.2
kafka-python==2.0.1